TMP: Dirty patch to force setting screen mode to 1280x800
In kms_screen_probe(), con->modes[0] (1920x1080) is used as
the screen mode by default and there is no way to override this.
Forcely use con->modes[2] (1280x800) as the screen mode to avoid
the stall in engine_load_config().
This may was required for CENTURY LCD-10000VH3 at least.

Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>